refactor(phase-c): consolidate Minecraft — features/minecraft/, integrations/, and extension split
Move all Minecraft-related code into coherent homes:

- src/minecraft.py -> features/minecraft/stats_sftp.py (pure SFTP stats helpers)
- src/minecraft_config.py -> features/minecraft/config.py (tuning knobs)
- src/minecraft_rcon.py -> src/integrations/minecraft_rcon.py (generic RCON
  socket client, belongs next to other 3rd-party protocol clients)
- extensions/minecraft.py -> extensions/minecraft/ package, split along the
  vlrgg/zevent template:
  * _common.py — Pydantic config schema, module constants, logger
  * status.py  — StatusMixin: 30s status poll, embed builders, scheduled event
  * stats.py   — StatsMixin: hourly stats task, table render, image cache
  * __init__.py — Minecraft(Extension) composition + setup(bot)

features/minecraft/__init__.py re-exports the public names so call sites can
keep using the top-level feature module.

Import updates:
- extensions/minecraft/_common.py, stats.py, status.py, __init__.py use
  features.minecraft and src.integrations.minecraft_rcon
- extensions/minecraft.py's three deferred imports (stats_cache,
  get_minecraft_stats_with_retry, get_config) are rewired to features.minecraft.

src/minecraft_rcon.py is currently unimported; moving it to src/integrations/
keeps the API surface stable for the future RCON wiring.
