+++ features/minecraft/__init__.py
@@ -1,0 +1,45 @@
+"""Minecraft feature layer: SFTP-based player stats and tuning knobs."""
+
+from features.minecraft.config import (
+    MINECRAFT_OPTIMIZATION_CONFIG,
+    apply_optimization_config,
+    get_config,
+)
+from features.minecraft.stats_sftp import (
+    MinecraftStatsCache,
+    calculate_level,
+    create_optimized_sftp_connection,
+    create_sftp_connection,
+    format_number,
+    get_all_player_stats_optimized,
+    get_minecraft_stats_with_retry,
+    get_player_stats,
+    get_player_stats_optimized,
+    get_single_player_stats_with_semaphore,
+    get_users,
+    read_nbt_file,
+    read_stats_file,
+    stats_cache,
+    ticks_to_hms,
+)
+
+__all__ = [
+    "MINECRAFT_OPTIMIZATION_CONFIG",
+    "MinecraftStatsCache",
+    "apply_optimization_config",
+    "calculate_level",
+    "create_optimized_sftp_connection",
+    "create_sftp_connection",
+    "format_number",
+    "get_all_player_stats_optimized",
+    "get_config",
+    "get_minecraft_stats_with_retry",
+    "get_player_stats",
+    "get_player_stats_optimized",
+    "get_single_player_stats_with_semaphore",
+    "get_users",
+    "read_nbt_file",
+    "read_stats_file",
+    "stats_cache",
+    "ticks_to_hms",
+]
